Add ProfileUserHeader and ProfileUserInfo Component
